Rename primitive package and add bounds struct
