+++ pkg/geo/geo.go
@@ -1,0 +1,5 @@
+// Package geo provides basic geometric and linear-algebraic primitives.
+package geo
+
+// epsilon serves as our minimum distance unit and floating-point epsilon.
+const epsilon = 1e-10
